there wasn't any reason to have 16 bit setpoint and measurement

+++ examples/DCMotorControllerWithAutoTune/DCMotorControllerWithAutoTune.ino
@@ -4,7 +4,6 @@
 #include <ArduinoPID.h>
 #include <AutoTuner.h>
 #include <PIDGains.h>
-#include <EEPROM.h>
 #include <util/atomic.h>
 
 
@@ -71,8 +70,8 @@ const int ONBOARD_LED =	  13;
  *---------------------------------------------------------------------
  */
 
-volatile int16_t positionMeasurement = 0;
-volatile int16_t positionSetpoint = 0; 
+volatile int32_t positionMeasurement = 0;
+volatile int32_t positionSetpoint = 0; 
 uint32_t pidExecutionTimer = 0;
 uint32_t pidExecutionIntervalUs = 1000000 / PID_FREQ_HZ;
 
@@ -260,7 +259,7 @@ void setup() {
  *---------------------------------------------------------------------
  */
 void loop() {
-  int16_t setpoint, measurement;
+  int32_t setpoint, measurement;
   uint32_t now = micros();
 
   if (now - pidExecutionTimer > pidExecutionIntervalUs){
